% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab6elife599carde734-package.R
\docType{package}
\name{lab6elife599carde734-package}
\alias{lab6elife599carde734}
\alias{lab6elife599carde734-package}
\title{lab6elife599carde734: Solutions to the knapsack problem}
\description{
Package for assignment 6 in Advanced R. It studies the knapsack problem, implementing different solutions for it with different complexities.
}
\details{
This package is the solution for Lab 6 in Advanced R.
TODO add description here
}
\author{
\strong{Maintainer}: Elisio Quintino \email{elife599@student.liu.se}

Authors:
\itemize{
  \item Carolina Rosario \email{carde734@student.liu.se}
}

}
\keyword{internal}
